python/pyproject.toml
[project]
name = "acp-sdk"
version = "1.0.0rc1"
description = "Agent Communication Protocol SDK"
readme = "README.md"
authors = []
requires-python = ">=3.13, <4.0"
dependencies = ["opentelemetry-api>=1.31.1", "pydantic>=2.11.1"]

[project.optional-dependencies]
client = [
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "opentelemetry-instrumentation-httpx>=0.52b1",
]
server = [
    "fastapi[standard]>=0.115.8",
    "opentelemetry-exporter-otlp-proto-http>=1.31.1",
    "opentelemetry-instrumentation-fastapi>=0.52b1",
    "opentelemetry-sdk>=1.31.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
